==> 3_modelling/5_train_valid_test_split.ipynb <==
import pandas as pd 
import pickle
import numpy as np
import random 
from sklearn.preprocessing import OrdinalEncoder
from sklearn.model_selection import train_test_split 
SEED=42
random.seed(SEED)
np.random.seed(seed=SEED)
pd.set_option('display.max_columns', None)

### Prep Train Valid and Test

df_data = pd.read_csv('../data/sale_data/2_sale_enriched.csv')
print(df_data.shape)
df_data.head(2)

df_data['prezzo'].quantile(q=[i/10 for i in range(1,10)])

bins = [0, 200000, 350000, 500000,700000,max(df_data['prezzo'])]
df_data['prezzo_binned'] = pd.cut(df_data['prezzo'], bins)

df_data['prezzo_binned'].value_counts()

def determine_optimal_split(df_to_split, train_size):
    grouped_df = df_to_split.groupby('train_test_split_key')[['house_id']].count().reset_index()
    grouped_df = grouped_df[grouped_df['house_id']>=2]
    filtered_df = df_to_split[df_to_split['train_test_split_key'].isin(grouped_df['train_test_split_key'])].sample(frac=1)
    df1, df2 = train_test_split(filtered_df, train_size=train_size, stratify=filtered_df['train_test_split_key'])
    df1 = pd.concat([df1, df_to_split[~df_to_split['train_test_split_key'].isin(grouped_df['train_test_split_key'])]])
    return df1, df2

splitting_cols = ['prezzo_binned', 'regione', 'stanze', 'bagni', 'tipologia_casa', 'classe_casa', 'tipologia_proprietà']
df_data['train_test_split_key'] = df_data[splitting_cols].apply(lambda row: '-'.join(row.values.astype(str)),axis=1)
train, test = determine_optimal_split(df_data, train_size=0.65)
valid, test = determine_optimal_split(test, train_size=0.5)
print('Training Observations:', train.shape[0], f'---> {round(100*train.shape[0]/df_data.shape[0],2)}%')
print('Validation Observations:', valid.shape[0], f'---> {round(100*valid.shape[0]/df_data.shape[0],2)}%')
print('Test Observations:', test.shape[0], f'---> {round(100*test.shape[0]/df_data.shape[0],2)}%')

categorical_features = ['bagni','stanze','disponibilità','cucina','tipologia_casa','classe_casa','tipologia_proprietà', 'giardino','infissi']

categorical_encoder = OrdinalEncoder()
train[categorical_features] = categorical_encoder.fit_transform(train[categorical_features])
valid[categorical_features] = categorical_encoder.transform(valid[categorical_features])
test[categorical_features] = categorical_encoder.transform(test[categorical_features])
with open('artifacts/categorical_encoder.pkl','wb') as f:
    pickle.dump(categorical_encoder, f)

train.to_csv('../data/modelling_data/train.csv', index=False)
valid.to_csv('../data/modelling_data/valid.csv', index=False)
test.to_csv('../data/modelling_data/test.csv', index=False)
