==> 3_modelling/8_refit.ipynb <==
import pandas as pd 
import pickle 
import lightgbm
import numpy as np
from sklearn.metrics import root_mean_squared_error, r2_score, d2_pinball_score,mean_absolute_error, mean_absolute_percentage_error


### Refit on Train + Val and Eval on Test

with open('artifacts/lightgbm_reg/quantile_lightgbm.pkl', 'rb') as fp:
    quantile_lightgbm = pickle.load(fp)
    
train = pd.read_csv('../data/modelling_data/train.csv')
valid = pd.read_csv('../data/modelling_data/valid.csv')
train_valid = pd.concat([train, valid], axis=0).sample(frac=1)
test = pd.read_csv('../data/modelling_data/test.csv')

features = train.select_dtypes(include=[np.number, bool]).columns
features = [i for i in features if i not in ['prezzo', 'log_prezzo', 'log_m2'] and 'emb_' not in i]
target='prezzo'

model_05 = lightgbm.LGBMRegressor(**quantile_lightgbm['lightgbm_q0.05']['study'].best_params)
model_05.fit(train_valid[features], train_valid[target])

model_50 = lightgbm.LGBMRegressor(**quantile_lightgbm['lightgbm_q0.5']['study'].best_params)
model_50.fit(train_valid[features], train_valid[target])

model_95 = lightgbm.LGBMRegressor(**quantile_lightgbm['lightgbm_q0.95']['study'].best_params)
model_95.fit(train_valid[features], train_valid[target])

def get_evaluation_metrics(y_true, y_pred, alpha=0.5):
    print('RMSE:', round(root_mean_squared_error(y_true=y_true, y_pred=y_pred),2))
    print('MAE:', round(mean_absolute_error(y_true=y_true, y_pred=y_pred),2))
    print('R2:', f'{round(100*r2_score(y_true=y_true, y_pred=y_pred),2)}%')
    print('Pinball Score', f'{round(100*d2_pinball_score(y_true=y_true, y_pred=y_pred, alpha=alpha),2)}%')
    print('MAPE:', f'{round(100*mean_absolute_percentage_error(y_true=y_true, y_pred=y_pred),2)}%')
    
def coverage_fraction(y, y_low, y_high):
    print('Coverage Fraction:', f'{round(100*np.mean(np.logical_and(y >= y_low, y <= y_high)),2)}%')

q_05_lightgbm_pred = model_05.predict(test[features])
q_50_lightgbm_pred = model_50.predict(test[features])
q_95_lightgbm_pred = model_95.predict(test[features])

get_evaluation_metrics(test[target], q_50_lightgbm_pred)

print('\nOther Quantiles')
print('Pinball Score Q05', f'{round(100*d2_pinball_score(y_true=test[target], y_pred=q_05_lightgbm_pred, alpha=0.05),2)}%')
print('Pinball Score Q95', f'{round(100*d2_pinball_score(y_true=test[target], y_pred=q_95_lightgbm_pred, alpha=0.95),2)}%')
coverage_fraction(test[target], q_05_lightgbm_pred, q_95_lightgbm_pred)

### Final refit

train_valid_test = pd.concat([train,valid,test],axis=0).sample(frac=1)

model_05 = lightgbm.LGBMRegressor(**quantile_lightgbm['lightgbm_q0.05']['study'].best_params)
model_05.fit(train_valid_test[features], train_valid_test[target])

model_50 = lightgbm.LGBMRegressor(**quantile_lightgbm['lightgbm_q0.5']['study'].best_params)
model_50.fit(train_valid_test[features], train_valid_test[target])

model_95 = lightgbm.LGBMRegressor(**quantile_lightgbm['lightgbm_q0.95']['study'].best_params)
model_95.fit(train_valid_test[features], train_valid_test[target])

with open('artifacts/refitted_lightgbm/model_05.pkl','wb') as f:
    pickle.dump(model_05, f)
    
with open('artifacts/refitted_lightgbm/model_50.pkl','wb') as f:
    pickle.dump(model_50, f)
    
with open('artifacts/refitted_lightgbm/model_95.pkl','wb') as f:
    pickle.dump(model_95, f)
